Split marking and tooltip layer
